+++ pkg/state/errors.go
@@ -25,9 +25,6 @@ var errCommandPrepareArgsNotSupportedByState = fmt.Errorf("%w: CommandPrepareArg
 var errQueryPrepareArgsNotSupportedByState = fmt.Errorf("%w: QueryPrepareArgs available for queries only", errors.ErrUnsupported)
 var errQueryCallbackNotSupportedByState = fmt.Errorf("%w: QueryCallback available for queries only", errors.ErrUnsupported)
 var errTest = errors.New("test")
-var errCurrentValueIsNotAnArray = errors.New("current value is not an array")
-var errFieldByIndexIsNotAnObjectOrArray = errors.New("field by index is not an object or array")
-var errNotImplemented = errors.New("not implemented")
 var errEntityRequiredForValueBuilder = errors.New("entity required for ValueBuilder")
 var errWorkspaceDescriptorNotFound = errors.New("WorkspaceDescriptor not found in workspace")
 var errDescriptorForUndefinedWorkspace = errors.New("workspace descriptor for undefined workspace")
@@ -37,54 +34,6 @@ var ErrQNameIsNotDefinedInWorkspace = errors.New("qname is not defined in worksp
 
 func errUnexpectedType(actual interface{}) error {
 	return fmt.Errorf("unexpected type: %v", actual)
-}
-
-func errInt32FieldUndefined(name string) error {
-	return errors.New("undefined int32 field: " + name)
-}
-
-func errInt64FieldUndefined(name string) error {
-	return errors.New("undefined int64 field: " + name)
-}
-
-func errFloat32FieldUndefined(name string) error {
-	return errors.New("undefined float32 field: " + name)
-}
-
-func errFloat64FieldUndefined(name string) error {
-	return errors.New("undefined float64 field: " + name)
-}
-
-func errBoolFieldUndefined(name string) error {
-	return errors.New("undefined bool field: " + name)
-}
-
-func errStringFieldUndefined(name string) error {
-	return errors.New("undefined string field: " + name)
-}
-
-func errBytesFieldUndefined(name string) error {
-	return errors.New("undefined bytes field: " + name)
-}
-
-func errQNameFieldUndefined(name string) error {
-	return errors.New("undefined QName field: " + name)
-}
-
-func errRecordIDFieldUndefined(name string) error {
-	return errors.New("undefined RecordID field: " + name)
-}
-
-func errNumberFieldUndefined(name string) error {
-	return errors.New("undefined number field: " + name)
-}
-
-func errCharsFieldUndefined(name string) error {
-	return errors.New("undefined chars field: " + name)
-}
-
-func errValueFieldUndefined(name string) error {
-	return errors.New("undefined value field: " + name)
 }
 
 func errIndexOutOfBounds(index int) error {
