--- pkg/state/provide.go
@@ -1,18 +1,0 @@
-/*
- * Copyright (c) 2022-present unTill Pro, Ltd.
- */
-
-package state
-
-func ProvideCommandProcessorStateFactory() CommandProcessorStateFactory {
-	return implProvideCommandProcessorState
-}
-func ProvideSyncActualizerStateFactory() SyncActualizerStateFactory {
-	return implProvideSyncActualizerState
-}
-func ProvideQueryProcessorStateFactory() QueryProcessorStateFactory {
-	return implProvideQueryProcessorState
-}
-func ProvideAsyncActualizerStateFactory() AsyncActualizerStateFactory {
-	return implProvideAsyncActualizerState
-}
