+++ pkg/state/stateprovide/impl_bundled_host_state_test.go
@@ -2,7 +2,7 @@
  * Copyright (c) 2022-present unTill Pro, Ltd.
  */
 
-package state
+package stateprovide
 
 import (
 	"context"
