+++ pkg/sys/storages/impl_app_secrets_storage_test.go
@@ -2,7 +2,7 @@
  * Copyright (c) 2022-present unTill Pro, Ltd.
  */
 
-package state
+package storages
 
 import (
 	"context"
@@ -11,6 +11,7 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/voedger/voedger/pkg/appdef"
 	"github.com/voedger/voedger/pkg/isecrets"
+	"github.com/voedger/voedger/pkg/state/stateprovide"
 	"github.com/voedger/voedger/pkg/sys"
 )
 
@@ -20,7 +21,7 @@ func TestAppSecretsStorage_BasicUsage(t *testing.T) {
 	secretBody := `{"secret":"key"}`
 	sr := &isecrets.SecretReaderMock{}
 	sr.On("ReadSecret", secret).Return([]byte(secretBody), nil)
-	s := ProvideAsyncActualizerStateFactory()(context.Background(), nilAppStructsFunc, nil, nil, nil, sr, nil, nil, nil, 0, 0)
+	s := stateprovide.ProvideAsyncActualizerStateFactory()(context.Background(), nilAppStructsFunc, nil, nil, nil, sr, nil, nil, nil, 0, 0)
 	kb, err := s.KeyBuilder(sys.Storage_AppSecret, appdef.NullQName)
 	require.NoError(err)
 	kb.PutString(sys.Storage_AppSecretField_Secret, secret)
@@ -32,7 +33,7 @@ func TestAppSecretsStorage_BasicUsage(t *testing.T) {
 }
 func TestAppSecretsStorage(t *testing.T) {
 	t.Run("Should return error when key invalid", func(t *testing.T) {
-		s := ProvideAsyncActualizerStateFactory()(context.Background(), nilAppStructsFunc, nil, nil, nil, nil, nil, nil, nil, 0, 0)
+		s := stateprovide.ProvideAsyncActualizerStateFactory()(context.Background(), nilAppStructsFunc, nil, nil, nil, nil, nil, nil, nil, 0, 0)
 		kb, err := s.KeyBuilder(sys.Storage_AppSecret, appdef.NullQName)
 		require.NoError(t, err)
