+++ pkg/sys/storages/impl_result_storage.go
